main-py/requirements.txt
llama-cpp-python
grpcio
grpcio-tools